Move DistanceGrid CUDA
