+++ applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaContactMapper.cu
@@ -34,7 +34,6 @@ namespace cuda
 
 extern "C"
 {
-    void RigidContactMapperCuda3f_setPoints2(unsigned int size, unsigned int nbTests, unsigned int maxPoints, const void* tests, const void* contacts, void* map);
     void SubsetContactMapperCuda3f_setPoints1(unsigned int size, unsigned int nbTests, unsigned int maxPoints, unsigned int nbPointsPerElem, const void* tests, const void* contacts, void* map);
 }
 
@@ -61,20 +60,6 @@ struct /*__align__(8)*/ GPUTestEntry
 
 __shared__ GPUTestEntry curTestEntry;
 
-__global__ void RigidContactMapperCuda3f_setPoints2_kernel(const GPUTestEntry* tests, const GPUContact* contacts, float3* map)
-{
-    if (threadIdx.x == 0)
-        curTestEntry = tests[blockIdx.x];
-
-    __syncthreads();
-
-    GPUContact c = contacts[curTestEntry.firstIndex + threadIdx.x];
-    if (threadIdx.x < curTestEntry.curSize)
-    {
-        map[curTestEntry.newIndex + threadIdx.x] = c.p2;
-    }
-}
-
 __global__ void SubsetContactMapperCuda3f_setPoints1_kernel(unsigned int nbPointsPerElem, const GPUTestEntry* tests, const GPUContact* contacts, int* map)
 {
     if (threadIdx.x == 0)
@@ -93,15 +78,6 @@ __global__ void SubsetContactMapperCuda3f_setPoints1_kernel(unsigned int nbPoint
 //////////////////////
 // CPU-side methods //
 //////////////////////
-
-void RigidContactMapperCuda3f_setPoints2(unsigned int size, unsigned int nbTests, unsigned int maxPoints, const void* tests, const void* contacts, void* map)
-{
-    // round up to 16
-    //maxPoints = (maxPoints+15)&-16;
-    dim3 threads(maxPoints,1);
-    dim3 grid(nbTests,1);
-    {RigidContactMapperCuda3f_setPoints2_kernel<<< grid, threads >>>((const GPUTestEntry*)tests, (GPUContact*)contacts, (float3*)map); mycudaDebugError("RigidContactMapperCuda3f_setPoints2_kernel");}
-}
 
 void SubsetContactMapperCuda3f_setPoints1(unsigned int size, unsigned int nbTests, unsigned int maxPoints, unsigned int nbPointsPerElem, const void* tests, const void* contacts, void* map)
 {
