+++ applications/plugins/SofaCUDA/src/SofaCUDA/component/collision/response/contact/CudaPenalityContactForceField.h
@@ -23,7 +23,7 @@
 
 #include <sofa/gpu/cuda/CudaTypes.h>
 #include <sofa/component/collision/response/contact/PenalityContactForceField.h>
-#include <sofa/gpu/cuda/CudaCollisionDetection.h>
+#include <sofa/gpu/cuda/GPUDetectionOutputVector.h>
 
 namespace sofa::component::collision::response::contact
 {
@@ -81,7 +81,7 @@ public:
 
     void addContact(int m1, int m2, const Deriv& norm, Real dist, Real ks, Real mu_s = 0.0f, Real mu_v = 0.0f, int oldIndex = 0);
 
-    void setContacts(Real distance, Real ks, sofa::core::collision::GPUDetectionOutputVector* outputs, bool useDistance, type::Mat3x3f* normXForm = NULL);
+    void setContacts(Real distance, Real ks, sofacuda::GPUDetectionOutputVector* outputs, bool useDistance, type::Mat3x3f* normXForm = nullptr);
 
     virtual void addForce(const core::MechanicalParams* mparams, DataVecDeriv& d_f1, DataVecDeriv& d_f2, const DataVecCoord& d_x1, const DataVecCoord& d_x2, const DataVecDeriv& d_v1, const DataVecDeriv& d_v2) override;
