+++ applications/plugins/SofaDistanceGrid/extensions/CUDA/src/SofaDistanceGrid/CUDA/CudaCollisionDetection.cu
@@ -21,7 +21,7 @@
 ******************************************************************************/
 #include <sofa/gpu/cuda/CudaCommon.h>
 #include <sofa/gpu/cuda/CudaMath.h>
-#include "mycuda.h"
+#include <sofa/gpu/cuda/mycuda.h>
 #include "cuda.h"
 
 #if defined(__cplusplus) && CUDA_VERSION < 2000
