+++ applications/plugins/SofaDistanceGrid/extensions/CUDA/src/SofaDistanceGrid/CUDA/config.h.in
@@ -19,57 +19,19 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#include <SofaDistanceGrid/initSofaDistanceGrid.h>
-#include "components/collision/DistanceGridCollisionModel.h"
-#include "components/forcefield/DistanceGridForceField.h"
-#include "RegisterModelToCollisionFactory.h"
+#pragma once
 
-namespace sofadistancegrid
+#include <sofa/config.h>
+
+#ifdef SOFA_BUILD_SOFADISTANCEGRID_CUDA
+#  define SOFA_TARGET @PROJECT_NAME@
+#  define SOFA_SOFADISTANCEGRID_CUDA_API SOFA_EXPORT_DYNAMIC_LIBRARY
+#else
+#  define SOFA_SOFADISTANCEGRID_CUDA_API SOFA_IMPORT_DYNAMIC_LIBRARY
+#endif
+
+namespace sofadistancegrid::cuda
 {
-extern "C" {
-SOFA_SOFADISTANCEGRID_API void initExternalModule();
-SOFA_SOFADISTANCEGRID_API const char* getModuleName();
-SOFA_SOFADISTANCEGRID_API const char* getModuleVersion();
-SOFA_SOFADISTANCEGRID_API const char* getModuleLicense();
-SOFA_SOFADISTANCEGRID_API const char* getModuleDescription();
-SOFA_SOFADISTANCEGRID_API const char* getModuleComponentList();
-}
-
-void initExternalModule()
-{
-    initSofaDistanceGrid();
-}
-
-void initSofaDistanceGrid()
-{
-    static bool first = true;
-    if (first)
-    {
-        first = false;
-    }
-    sofa::component::collision::registerDistanceGridCollisionModel();
-}
-
-const char* getModuleName()
-{
-    return sofadistancegrid::MODULE_NAME;
-}
-
-const char* getModuleVersion()
-{
-    return sofadistancegrid::MODULE_VERSION;
-}
-
-const char* getModuleLicense()
-{
-    return "LGPL";
-}
-
-const char* getModuleDescription()
-{
-    return "A distance grid stores the distance to an object into a 3d regular grid.  "
-           "This is an efficient data structure to get a distance approximation for   "
-           "point in space. This is why it is often used to implement collisions.     ";
-}
-
-} /// namespace sofadistancegrid
+    constexpr const char* MODULE_NAME = "@PROJECT_NAME@";
+    constexpr const char* MODULE_VERSION = "@PROJECT_VERSION@";
+} // namespace sofadistancegrid::cuda
