+++ applications/plugins/SofaDistanceGrid/extensions/CUDA/src/SofaDistanceGrid/CUDA/init.cpp
@@ -19,57 +19,44 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
+#include <SofaDistanceGrid/CUDA/init.h>
 #include <SofaDistanceGrid/initSofaDistanceGrid.h>
-#include "components/collision/DistanceGridCollisionModel.h"
-#include "components/forcefield/DistanceGridForceField.h"
-#include "RegisterModelToCollisionFactory.h"
+#include <SofaCUDA/init.h>
 
-namespace sofadistancegrid
+namespace sofadistancegrid::cuda
 {
+
 extern "C" {
-SOFA_SOFADISTANCEGRID_API void initExternalModule();
-SOFA_SOFADISTANCEGRID_API const char* getModuleName();
-SOFA_SOFADISTANCEGRID_API const char* getModuleVersion();
-SOFA_SOFADISTANCEGRID_API const char* getModuleLicense();
-SOFA_SOFADISTANCEGRID_API const char* getModuleDescription();
-SOFA_SOFADISTANCEGRID_API const char* getModuleComponentList();
+    SOFA_EXPORT_DYNAMIC_LIBRARY void initExternalModule();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleName();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleVersion();
+    SOFA_EXPORT_DYNAMIC_LIBRARY const char* getModuleComponentList();
 }
 
 void initExternalModule()
 {
-    initSofaDistanceGrid();
-}
-
-void initSofaDistanceGrid()
-{
-    static bool first = true;
-    if (first)
-    {
-        first = false;
-    }
-    sofa::component::collision::registerDistanceGridCollisionModel();
+    init();
 }
 
 const char* getModuleName()
 {
-    return sofadistancegrid::MODULE_NAME;
+    return MODULE_NAME;
 }
 
 const char* getModuleVersion()
 {
-    return sofadistancegrid::MODULE_VERSION;
+    return MODULE_VERSION;
 }
 
-const char* getModuleLicense()
+void init()
 {
-    return "LGPL";
+    static bool first = true;
+    if (first)
+    {
+        sofadistancegrid::initSofaDistanceGrid();
+        sofa::gpu::cuda::init();
+        first = false;
+    }
 }
 
-const char* getModuleDescription()
-{
-    return "A distance grid stores the distance to an object into a 3d regular grid.  "
-           "This is an efficient data structure to get a distance approximation for   "
-           "point in space. This is why it is often used to implement collisions.     ";
-}
-
-} /// namespace sofadistancegrid
+} // namespace volumetricrendering::cuda
