+++ applications/plugins/SofaDistanceGrid/src/SofaDistanceGrid/config.h.in
@@ -28,10 +28,16 @@
 #cmakedefine01 SOFADISTANCEGRID_HAVE_SOFA_GL
 
 #ifdef SOFA_BUILD_SOFADISTANCEGRID
-#  define SOFA_TARGET SofaDistanceGrid
+#  define SOFA_TARGET @PROJECT_NAME@
 #  define SOFA_SOFADISTANCEGRID_API SOFA_EXPORT_DYNAMIC_LIBRARY
 #else
 #  define SOFA_SOFADISTANCEGRID_API SOFA_IMPORT_DYNAMIC_LIBRARY
 #endif
+
+namespace sofadistancegrid
+{
+    constexpr const char* MODULE_NAME = "@PROJECT_NAME@";
+    constexpr const char* MODULE_VERSION = "@PROJECT_VERSION@";
+}
 
 #endif
